src: extend uv_queue_work to be thread-safe
Allow uv_queue_work() to be called from any thread, along with adding
two new API variations:

First, pass in only the work_cb so it is executed on the threadpool, but
no additional callbacks are called from the event loop thread. Before
the work_cb is called the req is removed from the internal queue so the
user is able to either free or reuse the req again. If after_work_cb is
NULL then loop must also be NULL or UV_EINVAL is returned. This cannot
be cancelled because uv_cancel() is not thread-safe, and there's no
thread available to the user they can use to cancel the request.

Second, pass in only the after_work_cb. This allows a callback to be
executed from the thread corresponding to the passed in loop. This req
cannot be cancelled because of how the internal mechanism to cancelling
is setup.