src: make req counter atomic
Switch active_reqs.count to use atomics so that they can be
registered/unregistered from any thread. This will be helpful to make
uv_queue_work() thread-safe.

While active_reqs.count is an unsigned int, the atomic operations are
int. This is because Windows only supports a LONG for its atomic
operations. The consequence of this is that the number of active request
is limited to 2^31 - 1 active reqs instead of 2^32. Though in practice
this shouldn't matter.