build: add ClusterFuzzLite setup
The goal is to enable continuous fuzzing of libuv. There is not a lot of
parsing or complex data handling in libuv so I think integrating with
OSS-Fuzz is too much. However, ClusterFuzzLite can be used to run
fuzzers for a small amount of seconds for each PR to ensure nothing
breaks. This commit adds ClusterFuzzLite setup as well as a fuzzer targeting
various operations. The fuzzer can be extended, e.g. it would be nice to
have more complex FS fuzzing, but I thought I would keep it as is for
now and see if you're interested in having fuzzing.

Signed-off-by: David Korczynski <david@adalogics.com>