macos: use absolute timeouts for cvar timed wait
Before this commit, libuv used pthread_cond_timedwait_relative_np() on
macOS out of necessity, because that platform did not support absolute
timeouts back when uv_cond_timedwait() was added.

Absolute timeouts are oblivious to preemption. Relative timeouts on the
other hand suffer from skew when preempted between computing the timeout
and waiting, because the thread ends up waiting longer than it needs to.

Libuv's API uses relative timeouts and is therefore not immune to that
issue but this commit should make it less bad and make libuv on macOS
align more closely with other platforms.

Fixes: https://github.com/libuv/libuv/issues/4372