src: add uv_credentials_changed() function
To be called by embedders to notify `libuv` that credentials of the
running process have changed. Only functional if iouring sqpoll-ring is
enabled: it synchronizes the ring credentials with the process'. In any
other scenario is a no-op.

Without this change, a privileged-dropped process could end up
performing privileged fs operations via the iouring.

Refs: https://github.com/advisories/GHSA-vr4q-vx84-9g5x