Reland "linux: eliminate a read on eventfd per wakeup"
Register the eventfd with EPOLLET to enable edge-triggered notification
where we're able to eliminate the overhead of reading the eventfd via
system call on each wakeup event.

When the eventfd counter reaches the maximum value of the unsigned
64-bit, which may not happen for the entire lifetime of the process, we
rewind the counter and retry.

This optimization saves one system call on each event-loop wakeup,
eliminating the overhead of `read(2)` as well as the extra latency for
each epoll wakeup.

Relands #4400