[pipe] use cached name for unix sockets
In order to unlink unix sockets, we already need to keep a copy of the path
in their handle struct. When getting the name of a pipe, we currently default
to icalling getsockname() on the underlying socket, which is notoriously
broken on some systems, notably FreeBSD and GNU/Hurd.

This change proposes to make use of the cached path whenever available.