linux: handle when libuv allocates a file descriptor normally used by stdio (#4558)
File descriptors 0 to 2 are used by stdio by default, but are not guaranteed
to be in use by stdio, in which case one or more of them may get used by libuv.

For example, STDERR can be closed by passing 2>&- on the program command line
in the shell. In which case, the file descriptor 2 would be used by libuv

libuv should handle this case without aborting