unix: fix return type for uv__udp_recvmmsg()
On FreeBSD (and other BSDs), recvmmsg() returns a ssize_t instead of an
int.  recvmsg_x() returns a ssize_t as well.  Avoid casting the return
value through an int as this might corrupt the value on big-endian
systems.

On Linux the return value will be sign-extended, and the sole caller
already saves it to a variable of type ssize_t, so this should have no
functional impact there.