linux: cap number of in-flight io_uring SQEs
Limit the number of in-flight submission entries to the size of the
completion queue. The kernel's sqpoll thread hoovers them up really
fast and we may end up posting way more submissions than fit in the
completion queue.

That's not fatal as such, the kernel stores them in an overflow queue
that we flush eventually, but that queue can grow pretty big if the
program submits lots of file operations.

Salient questions are:

1. Is this change a bug fix? If yes, why? If not, why not?

2. Is this change a (de)optimization? Is the kernel's overflow queue
   more or less efficient than libuv's thread pool?

This commit removes the IORING_SQ_CQ_OVERFLOW handling that was added in
commit 30fc896c ("unix: handle CQ overflow in iou ring") from May 2023
because it is no longer needed.

Fixes: https://github.com/libuv/libuv/issues/4598