Write to UV_HANDLE_BLOCKING_WRITES streams in the threadpool
On Unix, if a uv_stream_t has UV_HANDLE_BLOCKING_WRITES set, either
because we were unable to put the file in O_NONBLOCK mode when the
stream was initialized, or because the stream has been shared with a
subprocess via UV_INHERIT_STREAM, uv_write can block.

The nicest solution would be to use platform-specific features to do
non-blocking writes to files regardless of the O_NONBLOCK state, but a
fallback is needed regardless.  This change implements the fallback by
doing all writes to blocking streams in the threadpool.

Note that this also tweaks uv_spawn slightly: we clear O_NONBLOCK on
libuv-owned streams shared using UV_INHERIT_STREAM, but not on files
shared with UV_INHERIT_FD.